feat(web): Directorate of immigration chat widget UI changes (#12518)
* Add text

* Add translations

* Add integration id to check

---------
