+++ apps/web/components/ChatPanel/WatsonChatPanel/index.ts
@@ -3,4 +3,4 @@ export const WatsonChatPanel = dynamic(() => import('./WatsonChatPanel'), {
   ssr: false,
 })
 
-export * from './utils'
+export * from './directorateOfImmigrationChatUtils'
