+++ apps/web/components/Organization/Wrapper/config.ts
@@ -71,7 +71,7 @@ export const watsonConfig: Record<
     // Útlendingastofnun - Organization
     // https://app.contentful.com/spaces/8k0h54kbe6bj/entries/77rXck3sISbMsUv7BO1PG2
     '77rXck3sISbMsUv7BO1PG2': {
-      integrationID: '89a03e83-5c73-4642-b5ba-cd3771ceca54',
+      integrationID: '53c6e788-8178-448d-94c3-f5d71ec3b80e',
       region: 'eu-gb',
       serviceInstanceID: 'bc3d8312-d862-4750-b8bf-529db282050a',
       showLauncher: false,
