+++ apps/web/components/ServiceWeb/config.ts
@@ -108,7 +108,7 @@ export const watsonConfig: Record<
       namespaceKey: 'default',
     },
     [Organization.DIRECTORATE_OF_IMMIGRATION]: {
-      integrationID: '89a03e83-5c73-4642-b5ba-cd3771ceca54',
+      integrationID: '53c6e788-8178-448d-94c3-f5d71ec3b80e',
       region: 'eu-gb',
       serviceInstanceID: 'bc3d8312-d862-4750-b8bf-529db282050a',
       showLauncher: false,
